Move EFS tests into e2e.go so that they can be imported from elsewhere
